+++ core/moengage_flutter/CHANGELOG.md
@@ -3,6 +3,7 @@
 # Next Release
 
 ## Next Version
+- Added support for multiple self-handled in-apps
 - Android
     - Bundling `moe-android-sdk` & `inapp` dependencies with the plugin
     - `moe-android-sdk` version updated to `13.05.00`