+++ core/moengage_flutter_android/CHANGELOG.md
@@ -3,9 +3,11 @@
 # Next Release Date
 
 ## Next Version
+
 - Bundling `moe-android-sdk` & `inapp` dependencies with the plugin 
 - `moe-android-sdk` version updated to `13.05.00`
 - `inapp` version updated to `8.6.0`
+- Added support for multiple self-handled in-apps
 
 # 07-08-2024